Make migration work on SQLite

+++ db/migrate/030_add_some_indices.rb
@@ -1,13 +1,17 @@
 class AddSomeIndices < ActiveRecord::Migration
     def self.up
-        execute 'create index users_lower_email_index on users(lower(email))'
+        if ActiveRecord::Base.connection.adapter_name == "PostgreSQL"
+            execute 'create index users_lower_email_index on users(lower(email))'
+        end
 
         add_index :info_requests, :created_at
         add_index :info_requests, :title # For checking duplicates at new request time
 
-        execute "create index public_bodies_url_short_name_index on public_bodies(regexp_replace(replace(lower(short_name), ' ', '-'), '[^a-z0-9_-]', '', 'g'))"
-        execute "create index public_body_versions_url_short_name_index on public_body_versions(regexp_replace(replace(lower(short_name), ' ', '-'), '[^a-z0-9_-]', '', 'g'))"
-        execute "create index users_url_name_index on users (regexp_replace(replace(lower(name), ' ', '-'), '[^a-z0-9_-]', '', 'g'))"
+        if ActiveRecord::Base.connection.adapter_name == "PostgreSQL"
+            execute "create index public_bodies_url_short_name_index on public_bodies(regexp_replace(replace(lower(short_name), ' ', '-'), '[^a-z0-9_-]', '', 'g'))"
+            execute "create index public_body_versions_url_short_name_index on public_body_versions(regexp_replace(replace(lower(short_name), ' ', '-'), '[^a-z0-9_-]', '', 'g'))"
+            execute "create index users_url_name_index on users (regexp_replace(replace(lower(name), ' ', '-'), '[^a-z0-9_-]', '', 'g'))"
+        end
 
         add_index :post_redirects, :email_token
         add_index :post_redirects, :token
@@ -15,14 +19,18 @@ class AddSomeIndices < ActiveRecord::Migration
     end
 
     def self.down
-        execute 'drop index users_lower_email_index'
+        if ActiveRecord::Base.connection.adapter_name == "PostgreSQL"
+            execute 'drop index users_lower_email_index'
+        end
 
         remove_index :info_requests, :created_at
         remove_index :info_requests, :title 
 
-        execute 'drop index users_url_name_index'
-        execute 'drop index public_bodies_url_short_name_index'
-        execute 'drop index public_body_versions_url_short_name_index'
+        if ActiveRecord::Base.connection.adapter_name == "PostgreSQL"
+            execute 'drop index users_url_name_index'
+            execute 'drop index public_bodies_url_short_name_index'
+            execute 'drop index public_body_versions_url_short_name_index'
+        end
 
         remove_index :post_redirects, :email_token
         remove_index :post_redirects, :token
